refactor(ui,kit): the code-land runtime moves into @vendoai/ui/kit and the kit package is deleted
@vendoai/kit shipped 7 modules (627 LOC) whose whole job was to sit next to
@vendoai/ui and re-export `@vendoai/ui/kit` verbatim, then add the provider, the
guarded query/action hooks, the $state binding and the reshape/aggregate
vocabulary on top. Every one of those needs the browser bundle the box already
loads, and the $state store was a wrapper around ui's own `useKeyedState` — so
the split bought a second package, a second build, a second vitest config and a
dependency edge, and bought nothing else.

The 7 modules move into packages/ui/src/kit/ and the barrel grows ~50 lines of
export. `kit/src/state.ts` (useVendoState) lands as `vendo-state.ts` because
ui's kit already owns a `state.ts` (useKeyedState) — the very store it wrapped.
No new exports subpath: `./kit` already existed, and eject-templates pins that
list.

Pre-1.0 hard cut: no alias package, no deprecation shim. `@vendoai/kit` ->
`@vendoai/ui/kit`, every export name unchanged.
