+++ packages/box-template/package.json
@@ -12,7 +12,7 @@
   },
   "dependencies": {
     "@vendoai/core": "workspace:*",
-    "@vendoai/kit": "workspace:*",
+    "@vendoai/ui": "workspace:*",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
     "zod": "3.25.76"
