--- packages/kit/package.json
@@ -1,67 +1,0 @@
-{
-  "name": "@vendoai/kit",
-  "version": "0.8.0",
-  "description": "The code-land runtime for Vendo apps: the Kit components, the reshape + aggregate vocabulary, and the guarded data/action hooks a generated app imports inside its box.",
-  "keywords": [
-    "ai",
-    "agent",
-    "genui",
-    "react",
-    "vendo"
-  ],
-  "license": "Apache-2.0",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/runvendo/vendo.git",
-    "directory": "packages/kit"
-  },
-  "homepage": "https://vendo.run",
-  "bugs": "https://github.com/runvendo/vendo/issues",
-  "engines": {
-    "node": ">=20"
-  },
-  "publishConfig": {
-    "access": "public"
-  },
-  "type": "module",
-  "sideEffects": false,
-  "main": "./dist/index.js",
-  "types": "./dist/index.d.ts",
-  "exports": {
-    ".": {
-      "types": "./dist/index.d.ts",
-      "default": "./dist/index.js"
-    }
-  },
-  "files": [
-    "dist",
-    "README.md"
-  ],
-  "scripts": {
-    "build": "tsc -p tsconfig.json",
-    "test": "vitest run",
-    "typecheck": "tsc -p tsconfig.json --noEmit && tsc -p tsconfig.test.json --noEmit",
-    "test:coverage": "vitest run --coverage"
-  },
-  "dependencies": {
-    "@vendoai/core": "workspace:*",
-    "@vendoai/ui": "workspace:*"
-  },
-  "peerDependencies": {
-    "react": "^18.0.0 || ^19.0.0",
-    "react-dom": "^18.0.0 || ^19.0.0",
-    "zod": ">=3.25.0 <5"
-  },
-  "devDependencies": {
-    "@testing-library/react": "^16.1.0",
-    "@types/node": "^22.0.0",
-    "@types/react": "^19.2.0",
-    "@types/react-dom": "^19.2.0",
-    "jsdom": "^25.0.0",
-    "react": "^19.0.0",
-    "react-dom": "^19.0.0",
-    "typescript": "^5.6.0",
-    "vitest": "^2.1.0",
-    "zod": "3.25.76"
-  }
-}
