+++ packages/ui/test/kit/action.test.tsx
@@ -1,9 +1,9 @@
 import { act, cleanup, render, screen, waitFor } from "@testing-library/react";
 import type { Json, ToolOutcome } from "@vendoai/core";
 import { afterEach, describe, expect, it, vi } from "vitest";
-import { VendoAppProvider } from "../src/app-context.js";
-import { useToolAction } from "../src/action.js";
-import { useToolQuery } from "../src/query.js";
+import { VendoAppProvider } from "../../src/kit/app-context.js";
+import { useToolAction } from "../../src/kit/action.js";
+import { useToolQuery } from "../../src/kit/query.js";
 import { startDoor, type Door } from "./wire-fixture.js";
 
 afterEach(cleanup);
