+++ packages/ui/test/kit/surface.test.ts
@@ -1,12 +1,12 @@
 import { describe, expect, it } from "vitest";
-import * as kit from "../src/index.js";
+import * as kit from "../../src/kit/index.js";
 
 /**
  * The barrel IS the product here: an unexported module is a module a generated
  * app cannot import. The template and the builder prompt are both written
  * against this list, so it is asserted, not assumed.
  */
-describe("@vendoai/kit's export surface", () => {
+describe("@vendoai/ui/kit's export surface", () => {
   it("exports the runtime the blueprint promises", () => {
     for (const name of [
       // the reshape + aggregate vocabulary
@@ -34,7 +34,7 @@ describe("@vendoai/kit's export surface", () => {
     }
   });
 
-  it("re-exports the Kit itself, so code-land renders the same components", () => {
+  it("ships the Kit itself, so code-land renders the same components", () => {
     for (const name of ["Stat", "DataTable", "LineChart", "Button", "fmt", "KIT_COMPONENTS", "useKeyedState"]) {
       expect(typeof kit[name as keyof typeof kit], name).not.toBe("undefined");
     }
