+++ packages/ui/test/kit/wire-fixture.ts
