+++ scripts/dependency-guard.mjs
@@ -69,11 +69,6 @@ const LAYERS = {
   // core's KnowledgeAdapter contract; core-only, like the other engine blocks
   "@vendoai/knowledge": ["@vendoai/core"],
   "@vendoai/automations": ["@vendoai/core", "@vendoai/apps"],
-  // the code-land runtime shim (blueprint §5.4): what a generated app imports
-  // inside its box — the Kit re-exported, the reshape/aggregate vocabulary
-  // delegated to core, and the guarded data/action hooks. core + ui only: it
-  // ships into a browser bundle, so it must never reach a server block.
-  "@vendoai/kit": ["@vendoai/core", "@vendoai/ui"],
   // the harness runtime (build contract 2026-07-30 §2): the second multi-block
   // package after the umbrella. It runs any Harness — building the Turn, mapping
   // the guard's outcomes, mirroring onto today's wire, and emitting hot-path
@@ -107,13 +102,14 @@ const LAYERS = {
   ],
   // the universal box app template (blueprint §11): what every generated app is
   // built FROM inside its box, baked once per Vendo release. Private, never
-  // published, and deliberately reaches ONLY the code-land shim — an app in a
-  // box must never be able to import a server block.
+  // published, and deliberately reaches ONLY the browser layer — it imports the
+  // code-land runtime through `@vendoai/ui/kit`, and an app in a box must never
+  // be able to import a server block.
   // core joins for the DECLARED port contract (VENDO_DEV_PORT): the host that
   // mints the preview URL and the template that binds the socket must read one
   // constant, and they sit in different layers. core is the contract layer, not a
   // server block, so this does not weaken "an app in a box imports no server".
-  "@vendoai/box-template": ["@vendoai/core", "@vendoai/kit"],
+  "@vendoai/box-template": ["@vendoai/core", "@vendoai/ui"],
   // the canonical umbrella is the only package allowed to depend on every block
   "@vendoai/vendo": "*",
   // the unscoped compatibility package is a thin alias of the canonical umbrella
