+++ turbo.json
@@ -6,9 +6,9 @@
     "test:ui": { "dependsOn": ["^build", "build"], "env": ["CI", "VITEST_MIN_FORKS", "VITEST_MAX_FORKS", "VITEST_MIN_THREADS", "VITEST_MAX_THREADS"] },
     "test:coverage": { "dependsOn": ["^build", "build"], "outputs": ["coverage/**"], "env": ["CI", "POSTGRES_URL", "VITEST_MIN_FORKS", "VITEST_MAX_FORKS", "VITEST_MIN_THREADS", "VITEST_MAX_THREADS"] },
     // box-template.test.ts provisions and vite-builds packages/box-template, which imports
-    // @vendoai/kit — but apps must NOT depend on the kit/ui layer (scripts/dependency-guard.mjs),
+    // @vendoai/ui/kit — but apps must NOT depend on the ui layer (scripts/dependency-guard.mjs),
     // so the ordering is declared here instead. Without it turbo is free to schedule this suite
-    // before @vendoai/kit:build and the template build dies on an unresolvable @vendoai/kit.
+    // before @vendoai/ui:build and the template build dies on an unresolvable @vendoai/ui.
     "@vendoai/apps#test": { "dependsOn": ["^build", "build", "@vendoai/box-template#build"], "env": ["CI", "POSTGRES_URL", "VITEST_MIN_FORKS", "VITEST_MAX_FORKS", "VITEST_MIN_THREADS", "VITEST_MAX_THREADS"] },
     "@vendoai/apps#test:coverage": { "dependsOn": ["^build", "build", "@vendoai/box-template#build"], "outputs": ["coverage/**"], "env": ["CI", "POSTGRES_URL", "VITEST_MIN_FORKS", "VITEST_MAX_FORKS", "VITEST_MIN_THREADS", "VITEST_MAX_THREADS"] },
     // packages/ui's prebuild generates src/tree/jail/runtime-bundle.gen.ts INTO
